Reuse initializeScopes
