+++ packages/babel-parser/src/parser/index.js
@@ -5,14 +5,7 @@ import type { File /*::, JSXOpeningElement */ } from "../types";
 import type { PluginList } from "../plugin-utils";
 import { getOptions } from "../options";
 import StatementParser from "./statement";
-import { SCOPE_PROGRAM } from "../util/scopeflags";
 import ScopeHandler from "../util/scope";
-import ClassScopeHandler from "../util/class-scope";
-import ExpressionScopeHandler from "../util/expression-scope";
-import ProductionParameterHandler, {
-  PARAM_AWAIT,
-  PARAM,
-} from "../util/production-parameter";
 
 export type PluginsMap = Map<string, { [string]: any }>;
 
@@ -28,14 +21,8 @@ export default class Parser extends StatementParser {
     options = getOptions(options);
     super(options, input);
 
-    const ScopeHandler = this.getScopeHandler();
-
     this.options = options;
-    this.inModule = this.options.sourceType === "module";
-    this.scope = new ScopeHandler(this.raise.bind(this), this.inModule);
-    this.prodParam = new ProductionParameterHandler();
-    this.classScope = new ClassScopeHandler(this.raise.bind(this));
-    this.expressionScope = new ExpressionScopeHandler(this.raise.bind(this));
+    this.initializeScopes();
     this.plugins = pluginsMap(this.options.plugins);
     this.filename = options.sourceFilename;
   }
@@ -46,12 +33,7 @@ export default class Parser extends StatementParser {
   }
 
   parse(): File {
-    let paramFlags = PARAM;
-    if (this.hasPlugin("topLevelAwait") && this.inModule) {
-      paramFlags |= PARAM_AWAIT;
-    }
-    this.scope.enter(SCOPE_PROGRAM);
-    this.prodParam.enter(paramFlags);
+    this.enterInitialScopes();
     const file = this.startNode();
     const program = this.startNode();
     this.nextToken();
